+++ src/output.h
@@ -24,7 +24,7 @@
 
 typedef struct bear_output_t bear_output_t;
 
-bear_output_t * bear_open_json_output(char const * file, bear_output_filter_t const *);
+bear_output_t * bear_open_json_output(char const * file, bear_output_filter_t *);
 
 void bear_append_json_output(bear_output_t * handle, bear_message_t const * e);
 void bear_close_json_output(bear_output_t * handle);